requirements_dev.txt
numpy
pandas
jupyterlab
flatten_json
boto3
python-dotenv
pyyaml